Add parameter handling for !cr add !define command
!cr [RuleNumber] can now be used to query a specific CR entry.

!define [word] can be used to look up that word in the CR glossary.

The CR entries and glossary are parsed from the official CR txt release.

The url to retrieve the official txt release of the CR has to be
configured as process.env.CR_ADDRESS.

Slight rework of server.js:

Instantiate each aggregator object only once and not on every command. Use
an object as a lookup table instead of the switch case. Pass command to
getContent so an aggregator can handle multiple commands.