Use different ipg for the !ipg command
Switch to the ipg used and maintained by MTG Familiar from the annotated
ipg on blogs.magicjudges.org. In addition to being somewhat less cluttered
it is also all on one page.

Download and parse the ipg once on startup and not on every query.

Seperate parsing and output-formatting into several different functions.

Add some tests for the ipg module.